Retyped a counter so had to muck with many beans.  Forced creation of layered directories to work.
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/BRANCHES/WCM-DEV2/root@2930 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/repo/avm/Repository.java
@@ -43,7 +43,7 @@ public interface Repository
      * Get the latest version id.
      * @return The latest version.
      */
-    public long getLatestVersion();
+    public int getLatestVersion();
     
     /**
      * Inform this Repository that the root node is new.
@@ -131,7 +131,7 @@ public interface Repository
      * Get the version ids that this Repository has.
      * @return A Set of Version IDs.
      */
-    public Set<Long> getVersions();
+    public Set<Integer> getVersions();
     
     /**
      * Get the data bean.
